--- css-file/src/main/java/org/spc/file/api/monitor.java
@@ -1,8 +1,0 @@
-package org.spc.file.api;
-
-public class monitor {
-
-    //磁盘的使用情况（未占用的多少，占用的多少，系统占用的多少） : 直接查FATorder的list长度
-
-
-}
